Ignore unknown codecs in hello_world
When running hello_world on a transport stream it failed because that
stream contained 4 elementary streams: one for video, one for audio, one
for teletext and one for EPG data (program guide). For the latter no
codec was found.

This change just continues if for a stream no codec is found.
Actually for this sample it would suffice if there is just a video
stream.